Fixed loading Properties by ajax request. Added new Sequencing to export. Added template files as resource dependency. Fixed import courses with resource dependency. Removed redundant columns in Course and Node tables.

+++ IUDICO.CourseManagement/Models/ManifestModels/Importer.cs
@@ -105,7 +105,10 @@ namespace IUDICO.CourseManagement.Models.ManifestModels
         {
             foreach (var file in resource.Files)
             {
-                File.Copy(Path.Combine(_CourseTempPath, file.Href), Path.Combine(_CoursePath, file.Href));
+                if (!File.Exists(Path.Combine(_CoursePath, file.Href)))
+                {
+                    File.Copy(Path.Combine(_CourseTempPath, file.Href), Path.Combine(_CoursePath, file.Href));
+                }
             }
 
             foreach (var dependency in resource.Dependencies)
